+++ src/.env.example
@@ -1,0 +1,2 @@
+VITE_SUPABASE_URL=https://tuproyecto.supabase.co
+VITE_SUPABASE_ANON_KEY=tu-anon-key
