--- src/main.jsx
@@ -1,23 +1,0 @@
-// src/main.jsx
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-import { AuthProvider } from './context/AuthContext';
-import './index.css';
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-
-// Crear una instancia del cliente de React Query
-const queryClient = new QueryClient();
-
-createRoot(document.getElementById('root')).render(
-	<React.StrictMode>
-		<QueryClientProvider client={queryClient}>
-			<AuthProvider>
-				<App />
-			</AuthProvider>
-			<ReactQueryDevtools initialIsOpen={false} />
-		</QueryClientProvider>
-	</React.StrictMode>,
-);
