+++ src/postcss.config.js
@@ -1,0 +1,6 @@
+import tailwindcss from '@tailwindcss/postcss';
+import autoprefixer from 'autoprefixer';
+
+export default {
+	plugins: [tailwindcss(), autoprefixer()],
+};
