+++ src/src/lib/utils/utils.ts
@@ -1,0 +1,8 @@
+import { clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+import type { ClassValue } from 'clsx';
+
+export function cn(...inputs: ClassValue[]) {
+	return twMerge(clsx(...inputs));
+}
