+++ src/vite.config.ts
@@ -1,0 +1,8 @@
+import tailwindcss from '@tailwindcss/vite';
+import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite';
+
+// https://vite.dev/config/
+export default defineConfig({
+	plugins: [tailwindcss(), react()],
+});
